chore(*): Bumps version to 0.9.0

+++ examples/as2bindle.rs
@@ -79,7 +79,7 @@ async fn main() {
         name: format!("{}", path.file_name().unwrap().to_string_lossy()),
         media_type: "application/wasm".to_owned(),
         sha256: sha,
-        size: md.len() as u64,
+        size: md.len(),
         ..bindle::Label::default()
     };
